+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Add mapExecute() helper to execute a task function on an iterable data's input.
+
 ## [4.0.13] - 2024-05-29
 
 ### Changed